+++ CHANGES.rst
@@ -64,7 +64,20 @@
 
 - The in-memory cache now uses a better approximation of a LRU
   algorithm with less overhead, so more data should fit in the same
-  size cache. (For best performance, CFFI should be installed.)
+  size cache. (For best performance, CFFI should be installed; a
+  warning is generated if that is not the case.)
+
+- Support a persistent on-disk cache. This can greatly speed up
+  application warmup after a restart (such as deploying new code).
+  Some synthetic benchmarks show an 8-10x improvement. See :issue:`92`
+  for a discussion, and see the options ``cache-local-dir`` and
+  ``cache-local-dir-count``.
+
+- The in-memory cache is now smart enough not to store compressed
+  objects that grow during compression, and it uses the same
+  compression markers as zc.zlibstorage to avoid double-compression.
+  It can also gracefully handle changes to the compression format in
+  persistent files.
 
 2.0.0b1 (2016-06-28)
 ====================